+++ packages/Webkul/Admin/src/DataGrids/Lead/LeadDataGrid.php
@@ -360,9 +360,7 @@ class LeadDataGrid extends DataGrid
      */
     protected function getCustomAttributes(): Collection
     {
-        static $attributes;
-
-        return $attributes ??= Attribute::query()
+        return Attribute::query()
             ->where('entity_type', 'leads')
             ->where('is_user_defined', 1)
             ->get();